requirements.txt
# === Core for Week7 QLoRA fine-tune ===
transformers>=4.41.0
datasets>=2.20.0
peft>=0.11.1
bitsandbytes>=0.43.0
accelerate>=0.31.0
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
sentencepiece>=0.2.0
pandas>=2.2,<3
